+++ examples/todo_flutter/lib/event_handlers/logout_handler.dart
@@ -2,8 +2,9 @@ import 'package:etos_flutter/etos_flutter.dart';
 import 'package:todo_flutter/events/logout_event.dart';
 import 'package:todo_flutter/state/app_state.dart';
 
-class LogoutHandler {
-  void call(LogoutEvent event, _, StateSetter<AppState> set) {
-    set(UnauthenticatedAppState());
+class LogoutHandler extends EventHandler<AppState, LogoutEvent> {
+  @override
+  void call(LogoutEvent event) {
+    setState(UnauthenticatedAppState());
   }
 }