+++ etos/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+- Move `setState` and `getState` to `EventHandler` methods.
+- Add `callWhen` to allow `EventHandlers` to only be called in certain conditions.
+- Add `dispatch` to `EventHandler` to easily dispatch new `Events`.
+- Remove option to use functions as `EventHandlers`.
+
 ## 0.0.2
 
 - Remove unneeded casts.